+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Allow more than 100 concurrent connections to the Realtime Database emulator.
+- Fixes incorrect `databaseURL` inside the Cloud Functions emulator for new projects (#2965).
 - Fixes function URLs when emulating namespaced/grouped Cloud Functions (#2966).